+++ schemaParts/mos6502/src/java/pko/KiCadLogicalSchemeSimulator/components/mos6502/queue/Request.java
@@ -4,7 +4,7 @@ import pko.KiCadLogicalSchemeSimulator.components.mos6502.core.Utils;
 public class Request {
     public int address = -1;
     public boolean read;
-    public long payload;
+    public int payload;
     public Callback callback;
     public Request next;
