+++ schemaParts/stateMachine/test/java/pko/KiCadLogicalSchemeSimulator/components/stateMachine/test/StateMachineTest.java
@@ -35,8 +35,8 @@ import pko.KiCadLogicalSchemeSimulator.test.schemaPartTester.NetTester;
 import pko.KiCadLogicalSchemeSimulator.tools.Utils;
 
 public class StateMachineTest extends NetTester {
-    long[] states =
-            new long[]{0b1111110, 0b0110000, 0b1101101, 0b1111001, 0b0110011, 0b1011011, 0b1011111, 0b1110000, 0b1111111, 0b1111011, 0b0001110, 0b0110111, 0b1100111,
+    int[] states =
+            new int[]{0b1111110, 0b0110000, 0b1101101, 0b1111001, 0b0110011, 0b1011011, 0b1011111, 0b1110000, 0b1111111, 0b1111011, 0b0001110, 0b0110111, 0b1100111,
                     0b1110111, 0b0000001, 0b0};
 
     @Override
@@ -51,7 +51,7 @@ public class StateMachineTest extends NetTester {
 
     @Test
     protected void testStates() {
-        long mask = Utils.getMaskForSize(7);
+        int mask = Utils.getMaskForSize(7);
         for (int i = 0; i < states.length; i++) {
             setBus("in", i);
             if (i > 0) {
